Fix staggering of chat when scroll to bottom
